feat(sourcepoint_unified_cmp_ios): add SourcepointUnifiedCmpData.swift for data handling refactor(sourcepoint_unified_cmp_ios): move utility functions and classes to Utils.swift for better code organization refactor(sourcepoint_unified_cmp_ios): remove redundant code from SourcepointUnifiedCmpPlugin.swift to improve readability and maintainability
