+++ packages/sourcepoint_unified_cmp_ios/ios/Classes/SourcepointUnifiedCmpPlugin.swift
@@ -2,91 +2,9 @@ import ConsentViewController
 import Flutter
 import UIKit
 
+
 // This extension of Error is required to do use FlutterError in any Swift code.
 extension FlutterError: Error {}
-
-class Completer<T> {
-    typealias CompletionHandler = (T) -> Void
-
-    private var completionHandler: CompletionHandler?
-
-    func complete(result: T) {
-        DispatchQueue.main.async {
-            self.completionHandler?(result)
-        }
-    }
-
-    func setCompletionHandler(_ handler: @escaping CompletionHandler) {
-        completionHandler = handler
-    }
-}
-
-enum NotImplementedError: Error {
-    case notImplemented
-}
-
-extension SPGDPRVendorGrant {
-    func toHostAPIPurposeGrants() -> HostAPIGDPRPurposeGrants {
-        return HostAPIGDPRPurposeGrants(granted: granted, purposeGrants: purposeGrants)
-    }
-}
-
-extension ConsentStatus {
-    func toHostAPIConsentStatus() -> HostAPIConsentStatus {
-        return HostAPIConsentStatus(
-            consentedAll: consentedAll,
-            consentedToAny: consentedToAny,
-            granularStatus: nil, ///granularStatus, /// FIXME: not public in the ios sdk
-            hasConsentData: nil, ///hasConsentData,  /// FIXME: not public in the ios sdk
-            rejectedAny: rejectedAny,
-            rejectedLI: rejectedLI,
-            legalBasisChanges: legalBasisChanges,
-            vendorListAdditions: vendorListAdditions
-        )
-    }
-}
-
-extension SPConsent {
-    func toHostAPIGDPRConsent() throws -> HostAPIGDPRConsent {
-        if let value = self as? SPConsent<SPGDPRConsent> {
-            let tcData: [String?: String?]? = value.consents!.tcfData?.dictionaryValue?.reduce(into: [:]) { result, pair in
-                // Convert the value to String? if possible
-                let value: String? = (pair.value as? String) ?? (pair.value as? CustomStringConvertible)?.description
-
-                // Insert into the new dictionary
-                result[pair.key] = value
-            }
-            
-            let grants: [String?: HostAPIGDPRPurposeGrants?]? = value.consents?.vendorGrants.reduce(into: [:]) { result, pair in
-                // Convert the value to String? if possible
-                let value: SPGDPRVendorGrant = pair.value as SPGDPRVendorGrant
-
-                // Insert into the new dictionary
-                result[pair.key] = value.toHostAPIPurposeGrants()
-            }
-
-            return HostAPIGDPRConsent(
-                uuid: value.consents!.uuid,
-                tcData: tcData,
-                grants: grants,
-                euconsent: value.consents!.euconsent,
-                acceptedCategories: value.consents!.acceptedCategories,
-                apply: value.consents!.applies,
-                consentStatus: value.consents!.consentStatus.toHostAPIConsentStatus()
-            )
-        } else {
-            throw NotImplementedError.notImplemented
-        }
-    }
-}
-
-extension SPUserData {
-    func toHostAPISPConsent() -> HostAPISPConsent {
-        return HostAPISPConsent(
-            gdpr: try? gdpr?.toHostAPIGDPRConsent()
-        )
-    }
-}
 
 public class SourcepointUnifiedCmpPlugin: UIViewController, FlutterPlugin, SourcepointUnifiedCmpHostApi {
     private var consentManager: SPSDK!
@@ -121,16 +39,6 @@ public class SourcepointUnifiedCmpPlugin: UIViewController, FlutterPlugin, Sourc
         let api: SourcepointUnifiedCmpHostApi & NSObjectProtocol = SourcepointUnifiedCmpPlugin()
         SourcepointUnifiedCmpHostApiSetup.setUp(binaryMessenger: messenger, api: api)
     }
-}
-
-func getTopMostViewController() -> UIViewController? {
-    var topMostViewController = UIApplication.shared.keyWindow?.rootViewController
-
-    while let presentedViewController = topMostViewController?.presentedViewController {
-        topMostViewController = presentedViewController
-    }
-
-    return topMostViewController
 }
 
 extension SourcepointUnifiedCmpPlugin: SPDelegate {
